+++ pandora/cost_volume_confidence/cpp/includes/risk.hpp
@@ -46,7 +46,7 @@ py::list compute_risk_and_sampled_risk(
     py::array_t<float> sampled_ambiguity,
     py::array_t<double> etas,
     int nbr_etas,
-    py::array_t<int64_t> grids,
+    py::array_t<int> grids,
     py::array_t<float> disparity_range,
     bool sample_risk
 );
